+++ app/controllers/sheets_controller.rb
@@ -4,12 +4,12 @@ class SheetsController < ApplicationController
 
   before_action :authenticate_user!, except: [ :survey, :submit_survey, :submit_public_survey ]
   before_action :set_viewable_project, only: [ :index, :show, :print, :file, :verification_report ]
-  before_action :set_editable_project_or_editable_site, only: [ :edit, :transfer, :double_data_entry, :transactions, :new, :create, :update, :destroy, :unlock ]
-  before_action :redirect_without_project, only: [ :index, :show, :print, :edit, :transfer, :double_data_entry, :verification_report, :transactions, :new, :create, :update, :destroy, :unlock, :file ]
+  before_action :set_editable_project_or_editable_site, only: [ :edit, :transfer, :move_to_event, :double_data_entry, :transactions, :new, :create, :update, :destroy, :unlock ]
+  before_action :redirect_without_project, only: [ :index, :show, :print, :edit, :transfer, :move_to_event, :double_data_entry, :verification_report, :transactions, :new, :create, :update, :destroy, :unlock, :file ]
   before_action :set_viewable_sheet, only: [ :show, :print, :file, :verification_report ]
-  before_action :set_editable_sheet, only: [ :edit, :transfer, :double_data_entry, :transactions, :update, :destroy, :unlock ]
-  before_action :redirect_without_sheet, only: [ :show, :print, :edit, :transfer, :double_data_entry, :verification_report, :transactions, :update, :destroy, :unlock, :file ]
-  before_action :redirect_with_locked_sheet, only: [ :edit, :transfer, :double_data_entry, :verification_report, :update, :destroy ]
+  before_action :set_editable_sheet, only: [ :edit, :transfer, :move_to_event, :double_data_entry, :transactions, :update, :destroy, :unlock ]
+  before_action :redirect_without_sheet, only: [ :show, :print, :edit, :transfer, :move_to_event, :double_data_entry, :verification_report, :transactions, :update, :destroy, :unlock, :file ]
+  before_action :redirect_with_locked_sheet, only: [ :edit, :transfer, :move_to_event, :double_data_entry, :verification_report, :update, :destroy ]
 
   # GET /sheets
   # GET /sheets.json
@@ -235,6 +235,15 @@ class SheetsController < ApplicationController
 
       SheetTransaction.save_sheet!(@sheet, { subject_id: subject.id, subject_event_id: nil, last_user_id: current_user.id, last_edited_at: Time.now }, { }, current_user, request.remote_ip, 'sheet_update')
       redirect_to [@project, @sheet], notice: notice
+    end
+  end
+
+  def move_to_event
+    if subject_event = @sheet.subject.subject_events.find_by_id(params[:subject_event_id])
+      SheetTransaction.save_sheet!(@sheet, { subject_event_id: subject_event.id, last_user_id: current_user.id, last_edited_at: Time.now }, { }, current_user, request.remote_ip, 'sheet_update')
+      @subject = @sheet.subject # For move_to_event.js.erb partial
+    else
+      render nothing: true
     end
   end